+++ R/asp_ratio.R
@@ -1,7 +1,7 @@
 #' Compute aspect ratio for spatial input
 #'
 #' @description
-#' Computes the aspect ratio as width divided by height, or columns divided by
+#' Compute the aspect ratio as width divided by height, or columns divided by
 #' rows.
 #'
 #' @param x A `SpatRaster` object, an `sf`/`sfc` object or a numeric vector of
@@ -9,9 +9,6 @@
 #'   [sf::st_bbox()].
 #'
 #' @return A numeric scalar giving the aspect ratio.
-#' @export
-#' @encoding UTF-8
-#' @keywords internal
 #' @examples
 #' \donttest{
 #' library(terra)
@@ -20,6 +17,9 @@
 #' plot(x)
 #' asp_ratio(x)
 #' }
+#' @export
+#' @encoding UTF-8
+#' @keywords internal
 asp_ratio <- function(x) {
   if (inherits(x, "SpatRaster")) {
     ratio <- terra::ncol(x) / terra::nrow(x)
@@ -30,7 +30,7 @@ asp_ratio <- function(x) {
     # Handle a bounding box in `xmin`, `ymin`, `xmax`, `ymax` order.
     ratio <- (x[3] - x[1]) / (x[4] - x[2])
   } else {
-    stop("Cannot compute the aspect ratio for this input.", call. = FALSE)
+    cli::cli_abort("Cannot compute the aspect ratio for {.arg x}.")
   }
 
   ratio <- as.double(ratio)