+++ tests/testthat/helpers.R
@@ -1,6 +1,3 @@
-# Use these helpers only for testing purposes.
-# Load a test raster from a temporary file.
-
 testhelp_load_rast <- function(path) {
   ext <- tools::file_ext(path)
   tmp <- tempfile(fileext = paste0(".", ext))