+++ tests/testthat/test-errors_messages.R
@@ -36,10 +36,7 @@ test_that("Message in lonlat sf", {
   )
   x <- sf::st_transform(x, 4326)
   img <- system.file("img/UK_flag.png", package = "rasterpic")
-  expect_message(
-    rasterpic_img(x, img),
-    "Input 'x' has geographic coordinates. Assuming planar coordinates."
-  )
+  expect_snapshot(s <- rasterpic_img(x, img), )
 
   x2 <- sf::st_transform(x, 3857)
 
@@ -50,10 +47,7 @@ test_that("Message in lonlat raster", {
   x <- testhelp_load_rast(system.file("tiff/elev.tiff", package = "rasterpic"))
   x <- terra::project(x, "epsg:4326")
   img <- system.file("img/UK_flag.png", package = "rasterpic")
-  expect_message(
-    rasterpic_img(x, img),
-    "Input 'x' has geographic coordinates. Assuming planar coordinates."
-  )
+  expect_snapshot(s <- rasterpic_img(x, img))
 
   x2 <- terra::project(x, "epsg:3857")
   expect_silent(rasterpic_img(x2, img))